added ability to flip on/off my “event loop” attempt. tl;dr: it doesn’t help
it does the SAME thing for Validate: ie: validate just hangs

for Purchase, it’s worse and just fails, because the thread spawned is not a “single thread apartment” type. maybe if we launch it from a STA thread it would work?  still not getting callbacks for started, err, end.

Signed-off-by: David M. Cotter <dave@kjams.com>